+++ src/components/Products/Subcategories/Subcategories.js
@@ -5,6 +5,7 @@ import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faPlus, faEdit, faTrash, faScrewdriverWrench } from '@fortawesome/free-solid-svg-icons';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import { useHistory } from 'react-router-dom';
+import Swal from 'sweetalert2';
 
 const Subcategories = ({ categoryId }) => {
   const [subcategories, setSubcategories] = useState([]);
@@ -64,16 +65,43 @@ const Subcategories = ({ categoryId }) => {
     }
   };
 
-  const handleDeleteSubcategory = async (subcategoryId) => {
-    try {
-      await axios.delete(`https://localhost:7028/api/categories/${categoryId}/subcategories/${subcategoryId}`);
+  const handleDeleteSubcategory = async (subcategoryId) => {  
+    Swal.fire({
+      title: '¿Estás seguro?',
+      text: '¡No podrás revertir esto!',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Sí, eliminarlo!'
+    }).then(async (result) => {
+      if (result.isConfirmed) {
+        try {
 
-      const updatedSubcategories = subcategories.filter((subcategory) => subcategory.subcategoryId !== subcategoryId);
-      setSubcategories(updatedSubcategories);
-    } catch (error) {
-      console.error('Error deleting subcategory:', error);
-    }
+          await axios.delete(`https://localhost:7028/api/categories/${categoryId}/subcategories/${subcategoryId}`);
+
+          const updatedSubcategories = subcategories.filter((subcategory) => subcategory.subcategoryId !== subcategoryId);
+          
+          setSubcategories(updatedSubcategories);
+          
+          Swal.fire(
+            '¡Eliminado!',
+            '¡Tu subcategoría ha sido eliminada.',
+            'success'
+          );
+
+        } catch (error) {
+          console.error('Error deleting subcategory:', error);
+          Swal.fire(
+            'Error',
+            'Hubo un problema al eliminar la subcategoría.',
+            'error'
+          );
+        }
+      }
+    });
   };
+  
 
   const handleShowCreateModal = () => {
     setModalAction('create');
