+++ src/components/Sales/Contributions/Contributions.js
@@ -5,7 +5,7 @@ import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faPlus, faEdit, faTrash, faEye } from '@fortawesome/free-solid-svg-icons';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import { useHistory } from 'react-router-dom';
-
+import Swal from 'sweetalert2';
 const Contributions = ({orderId}) => {
     
   
@@ -87,16 +87,44 @@ const Contributions = ({orderId}) => {
   };
 
   const handleDeleteContribution = async (contributionId) => {
-    try{
-        await axios.delete (`https://localhost:7028/api/orders/${orderId}/contributions/${contributionId}`);
+    
+    Swal.fire({
+      title: '¿Estás seguro?',
+      text: '¡No podrás revertir esto!',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Sí, eliminarlo!'
+    }).then(async (result) => {
+        
+      if (result.isConfirmed) {
+        try {
+         
+          await axios.delete(`https://localhost:7028/api/orders/${orderId}/contributions/${contributionId}`);
 
-        const updatedContributions = contributions.filter((contribution)=> contribution.contributionID !== contributionId);
-        setContributions(updatedContributions);
-    }catch (error){
-        console.error ('error deleting contribution', error);
+          const updatedContributions = contributions.filter((contribution) => contribution.contributionID !== contributionId);
 
-    }
+          setContributions(updatedContributions);
+          
+          Swal.fire(
+            '¡Eliminado!',
+            '¡Tu contribución ha sido eliminada.',
+            'success'
+          );
+        } catch (error) {
+          console.error('Error deleting contribution:', error);
+          
+          Swal.fire(
+            'Error',
+            'Hubo un problema al eliminar la contribución.',
+            'error'
+          );
+        }
+      }
+    });
   };
+  
 
   const handleShowCreateModal = () => {
     setModalAction('create');
